requirements.txt
azure-functions
pandas
xlrd
